requirements.txt
feedparser
python-dotenv
feedgen
flask
gunicorn
